Consolidate Error\BaseException and Error\Exception into a single class.
